+++ viewer/src/index.js
@@ -5,20 +5,13 @@ import App from './components/App.jsx';
 import configureStore from './store';
 import Actions from './actions/actions';
 import { loadFile } from './utils';
-
-const DjVu = window.DjVu;
-
-if (!DjVu) {
-    throw new Error("There is no DjVu object! You have to include the DjVu.js library first!");
-}
+import DjVu from './DjVu';
 
 const store = configureStore();
 
 DjVu.Viewer = {
-    VERSION: '0.1.6',
+    VERSION: '0.1.7',
     init(element) {
-        //element.style.width = window.innerWidth * 0.9 + 'px';
-        element.style.height = window.innerHeight * 0.95 + 'px';
         ReactDOM.render(
             <Provider store={store}>
                 <App />